split schema into multiple files
