+++ rust/util_libs/db/src/schemas/mod.rs
@@ -1,0 +1,11 @@
+pub mod metadata;
+pub mod developer;
+pub mod host;
+pub mod hoster;
+pub mod public_service;
+pub mod user_info;
+pub mod user;
+pub mod workload;
+
+/// Name of the main database for the Holo Hosting system
+pub const DATABASE_NAME: &str = "holo-hosting";
